Work from Chapter 2
Naive implementation of :verse

2.3 add second test

2.3 rename unused param to descriptive name

2.3 consume params in template to generalize method

2.5 add test for second verse

2.5 pass test for second verse

Prefers duplication over abstractions that make the situation less
clear. The alternative would be to use a :pluralize method on 'bottle',
but that could be a premature optimization. This code is pretty
straightforward and easy to understand, even with the duplication.

Rename second verse test

2.6 test verse 1

Convert if conditional to case/when

2.6 add verse 1 variant, pass :test_verse_1

2.6 test for verse 0 variation

2.6 add 0 verse case, pass :test_verse_0

2.7 first test for :verses

2.7 define Bottles#verses

2.7 fix Bottles#verses signature

2.7 shameless green for Bottles#verses

Add second test for :verses to catch edge cases of :verse

Fix typo in :verse(1) case test

Fix typo in :verse(1) case

2.7 use range and .downto to implement :verses

2.9 add test for whole song

Add signature for :song

2.9 Implement :song

... this test wasn't actually introduced, I just got overambitious

2.12 implement :song test

testing against the hardcoded output leaves this test uncoupled to
implementation details

Reorder methods in Bottles class

makes it more like the examples in the book, and is also the order in
which I expect consumers of this class to prefer them.